chanbuff IS used appropriately: in channelserve, data is read in to servetestbuff, and sprintf'd into chanbuff
No longer prints those extra messages.

By having the channel print the number of characters read in, we can confirm that after reading the data from client, channel keeps running.

HOWEVER, read() is supposedly blocking by default. And according to the printing of # of chars read in (continually printing zero), read is continuing to read stuff in from socket, even when nothing is being written to socket. Research indicates that this means the other side of the socket (ie: the client) has closed their connection to the socket.